requirements.txt
numpy>=1.20
matplotlib>=3.4
pyyaml>=6.0
viennarna>=2.5
customtkinter
Pillow
# Java 8+ is required for RNArtistCore (not installable via pip)
# Download from https://adoptium.net/ or https://www.java.com/ and ensure 'java' is in your PATH 